Phase 3 development and testing done

+++ shared/models/src/restaurant.rs
@@ -1,4 +1,5 @@
 use chrono::{DateTime, Utc};
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
 use uuid::Uuid;
@@ -25,7 +26,7 @@ pub struct FoodCategory {
     pub restaurant_id: Uuid,
     pub name: String,
     pub description: Option<String>,
-    pub display_order: i32,
+    pub display_order: i64,
     pub is_active: bool,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
@@ -39,7 +40,7 @@ pub struct Food {
     pub category_id: Uuid,
     pub name: String,
     pub description: Option<String>,
-    pub price: f64,
+    pub price: Decimal,
     pub is_available: bool,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
@@ -61,7 +62,7 @@ pub struct CreateFoodCategoryRequest {
     #[validate(length(min = 1, max = 255))]
     pub name: String,
     pub description: Option<String>,
-    pub display_order: Option<i32>,
+    pub display_order: Option<i64>,
 }
 
 /// Create food request
@@ -71,12 +72,11 @@ pub struct CreateFoodRequest {
     #[validate(length(min = 1, max = 255))]
     pub name: String,
     pub description: Option<String>,
-    #[validate(range(min = 0.01))]
-    pub price: f64,
+    pub price: Decimal,
 }
 
 /// Menu response (category with foods)
-#[derive(Debug, Serialize)]
+#[derive(Debug, Serialize, Deserialize)]
 pub struct MenuCategory {
     #[serde(flatten)]
     pub category: FoodCategory,
@@ -84,7 +84,7 @@ pub struct MenuCategory {
 }
 
 /// Full menu response
-#[derive(Debug, Serialize)]
+#[derive(Debug, Serialize, Deserialize)]
 pub struct MenuResponse {
     pub restaurant: Restaurant,
     pub categories: Vec<MenuCategory>,
